+++ src/features/profile/types/index.ts
@@ -14,7 +14,7 @@ export interface ProfilePageProps {
 }
 
 export interface ProfileEditData {
-  name: string;
+  nick?: string;
   description?: string;
   image?: string | null;
 }
@@ -33,25 +33,10 @@ export interface BookClub {
   memberCount: number;
   isLiked: boolean;
   isCanceled: boolean;
-  imageUrl: string | undefined;
+  imageUrl?: string | undefined;
   isPast: boolean;
   clubStatus: 'pending' | 'confirmed' | 'closed';
   reviewScore?: number | undefined;
-}
-
-export interface Review {
-  reviewId: number;
-  userId: number;
-  bookClubId: number;
-  clubName: string;
-  rating: number;
-  comment: string;
-  clubImgUrl?: string | undefined;
-  clubImgAlt?: string | undefined;
-  profileImg?: string;
-  userName?: string;
-  createdAt: string;
-  bookClubType: 'FREE' | 'FIXED';
 }
 
 export type orderType = 'DESC' | 'ASC' | 'END';
