+++ src/features/profile/types/profile.ts
@@ -1,0 +1,5 @@
+export interface editProfileParams {
+  nickname?: string;
+  image?: string;
+  description?: string;
+}
