+++ src/features/react-query/book-club/customHooks.ts
@@ -2,6 +2,8 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { BookClubForm } from '@/features/club-create/types';
 import { createBookClub } from '@/features/club-create/api';
 import { bookClubs } from './queries';
+// import { editProfile } from '@/features/profile/api/editProfileApi';
+// import { editProfileParams } from '@/features/profile/types/profile';
 
 export function useBookClubCreateMutation() {
   const queryClient = useQueryClient();
@@ -18,3 +20,12 @@ export function useBookClubCreateMutation() {
     },
   });
 }
+
+// export function useEditProfile() {
+//   const queryClient = useQueryClient();
+
+//   return useMutation({
+//     mutationFn: (data: editProfileParams) => editProfile(data),
+//     onSuccess: () => {},
+//   });
+// }
