+++ src/features/react-query/book-club/queries.ts
@@ -78,3 +78,5 @@ export const bookClubs = createQueryKeys('bookClubs', {
       }),
   }),
 });
+
+export const profile = createQueryKeys('profile', {});
