+++ src/types/review.ts
@@ -1,0 +1,14 @@
+export interface Review {
+  reviewId: number;
+  userId: number;
+  bookClubId: number;
+  clubName: string;
+  rating: number;
+  comment: string;
+  clubImgUrl?: string | undefined;
+  clubImgAlt?: string | undefined;
+  profileImg?: string;
+  userName?: string;
+  createdAt: string;
+  bookClubType: 'FREE' | 'FIXED';
+}
